+++ src/filter/filter-type.js
@@ -1,0 +1,14 @@
+/**
+ * @fileoverview Типы фильтров
+ * @author Taova
+ */
+'use strict';
+
+
+var filterType = {
+  'popular': 'filter-popular',
+  'date': 'filter-new',
+  'comments': 'filter-discussed'
+};
+
+module.exports = filterType;
