+++ src/picture/get-picture-elem.js
@@ -1,0 +1,40 @@
+/**
+ * @fileoverview Получение элемента изображения
+ * @author TAova
+ */
+'use strict';
+
+var elementToClone;
+var templateElement = document.querySelector('#picture-template');
+
+if ('content' in templateElement) {
+  elementToClone = templateElement.content.querySelector('.picture');
+} else {
+  elementToClone = templateElement.querySelector('.picture');
+}
+/**
+ * @param {Object} data
+ * @param {HTMLElement} container
+ * @return {HTMLElement}
+ */
+
+var getPicture = function(data) {
+  var element = elementToClone.cloneNode(true);
+  var elemImg = element.querySelector('img');
+  element.querySelector('.picture-comments').textContent = data.comments;
+  element.querySelector('.picture-likes').textContent = data.likes;
+
+  var elemImgData = new Image();
+  elemImgData.onload = function() {
+    elemImg.src = elemImgData.src;
+    elemImgData.width = 182;
+    elemImgData.height = 182;
+  };
+  elemImgData.onerror = function() {
+    element.classList.add('picture-load-failure');
+  };
+  elemImgData.src = data.url;
+  return element;
+};
+
+module.exports = getPicture;
